GI-56 Integrate email with devise for confirmation
This change integrates the notify mailer with devise by creating a
custom mailer. The flash message was not being passed to the page after
signing up, as this was redirecting through to the home page, then
redirecting back to the sign in page.  To fix this the registrations
controller in devise is overridden, and a new flash message is passed
in.