Add kubectl deployment files
This adds the initial deployment files required to make a manual
deployment to the Moj Cloud platform. This is a temporary solution until
the deployment is moved to the circle ci pipeline.

Also added:

Git crypt to enable secrets to be stored
Style changes to allow css to be precompiled
Rails config changes to create UAT environment
Changes to dockerfile to work with kubernetes