GI-79 Only admins can comment on not_proceeding
We want only admins to be able to comment on not proceeding ideas.
Refactored the .approved_by_admin? logic to a helper class, as it is
starting to get complex.  We can also use this helper class to introduce
further methods such as user_can_edit_comment for future stories.