GI-81 Fix comments showing user and idea ID
Changed comments to include users when getting records so that we can
show the user email address without querying the DB again.
Also removed idea from the comments table as you are already on the idea
when you view this.

Tests updated to reflect the comment must include a user as well.