GI-106 Separate seed files
Seed files separated so that we can have different production and non
production seeds.  If we need to separate further we can do simply by
modifying the seeds.rb file and add more seeds.

Further work needs to be done to create more realistic seeds but this is
not covered in this card.