GI-131 Validation for delete comments
Add validation for delete comments.  Users can delete their own comment,
and admin users can delete anyones comment.

When a comment is deleted it is actually redacted, so the page displays
"This comment has been deleted" but the comment still exists in the
database only with a redacted flag set to true.